services/auth/requirements.txt
fastapi==0.116.1
pydantic==2.11.7
pydantic-settings==2.10.1
uvicorn==0.35.0
uvloop==0.21.0

alembic==1.16.5
asyncpg==0.30.0
greenlet==3.1.1
sqlalchemy==2.0.43

aio-pika==9.5.7
httpx==0.28.1
PyJWT==2.10.1
tenacity==8.2.3

prometheus-fastapi-instrumentator==7.0.0

opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-sdk==1.22.0
